+++ src/main/java/com/team1/hrbank/domain/employee/constant/EmployeeErrorCode.java
@@ -1,0 +1,17 @@
+package com.team1.hrbank.domain.employee.constant;
+
+import com.team1.hrbank.global.constant.ErrorCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum EmployeeErrorCode implements ErrorCode {
+
+  EMPLOYEE_NOT_FOUND(404, "찾을 수 없습니다.", "해당 직원을 찾을 수 없습니다."),
+  EMPLOYEE_EMAIL_DUPLICATE(409, "중복된 이메일 입니다.", "해당 이메일은 중복된 이메일입니다.");
+
+  private final int status;
+  private final String message;
+  private final String details;
+}
